Refactor tests a bit in prep for new test files.
